Move route UI ownership to pages
