+++ src/features/deck-editor/index.ts
@@ -1,1 +1,3 @@
-export { DeckFormContainer } from "./containers/DeckFormContainer";
+export { DeckForm, type DeckFormFields, type DeckFormProps } from "./components/DeckForm";
+export { useDeckEditorActions } from "./hooks/useDeckEditorActions";
+export { deckFormSchema, type DeckFormValues } from "./lib/deckFormSchema";
