+++ src/features/settings/index.ts
@@ -1,1 +1,3 @@
-export { ConfigContainer } from "./containers/ConfigContainer";
+export { ConfigForm, type ConfigFormFields, type ConfigFormProps } from "./components/ConfigForm";
+export { useAccountOperations } from "./hooks/useAccountOperations";
+export { useConfigFormState } from "./hooks/useConfigFormState";
