+++ src/pages/settings/ui/SettingsView.spec.tsx
@@ -1,16 +1,11 @@
-/**
- * @file Verifies the "ConfigFormTemplate" contract with automated examples.
- * The examples make the expected behavior concrete with cases such as "composes the config form
- * under a compact page heading without a redundant surface".
- */
-
 import { render, screen } from "@testing-library/react";
 import { describe, expect, it } from "vitest";
 import "@testing-library/jest-dom/vitest";
 
-import type { ConfigFormFields } from "@/features/settings/components/ConfigForm";
-import { ConfigFormTemplate } from "@/features/settings/components/templates/ConfigFormTemplate";
+import type { ConfigFormFields } from "@/features/settings";
 import { createConfig } from "@/test/factories";
+
+import { SettingsView } from "./SettingsView";
 
 const fields: ConfigFormFields = {
   showHeader: { name: "showHeader" },
@@ -24,10 +19,10 @@ const fields: ConfigFormFields = {
   cardInterval: { name: "cardInterval", value: "5", onChange: () => undefined },
 };
 
-describe("ConfigFormTemplate", () => {
-  it("composes the config form under a compact page heading without a redundant surface", () => {
+describe("SettingsView", () => {
+  it("composes the config form under one page heading without the application shell", () => {
     render(
-      <ConfigFormTemplate
+      <SettingsView
         configForm={{
           config: createConfig(),
           fields,
@@ -37,10 +32,10 @@ describe("ConfigFormTemplate", () => {
       />
     );
 
-    const heading = screen.getByRole("heading", { level: 1, name: "Settings" });
     expect(screen.getAllByRole("heading", { level: 1 })).toHaveLength(1);
-    expect(heading).toHaveClass("text-title");
+    expect(screen.getByRole("heading", { level: 1, name: "Settings" })).toHaveClass("text-title");
     expect(screen.getByText("Changes are saved automatically")).toHaveClass("text-ink-muted");
-    expect(screen.getByRole("region", { name: "Account" })).toBeInTheDocument();
+    expect(screen.getByRole("region", { name: "Account" })).toBeVisible();
+    expect(screen.queryByRole("button", { name: "tango" })).not.toBeInTheDocument();
   });
 });
