+++ src/shared/ui/layout/index.ts
@@ -1,2 +1,1 @@
 export { Layout } from "./Layout";
-export type { LayoutProps } from "./Layout";
